Add column containing number of missing data values

It can be useful when doing data analysis to add the number of missing data points into your dataframe. add_n_miss adds a column named "n_miss", which contains the number of missing values in that row.

Usage

add_n_miss(data, ..., label = "n_miss")

Arguments

data

a dataframe

...

Variable names to use instead of the whole dataset. By default this looks at the whole dataset. Otherwise, this is one or more unquoted expressions separated by commas. These also respect the dplyr verbs starts_with, contains, ends_with, etc. By default will add "_all" to the label if left blank, otherwise will add "_vars" to distinguish that it has not been used on all of the variables.

label

character default is "n_miss".

Value

a dataframe

See also

Examples

airquality %>% add_n_miss()
#>     Ozone Solar.R Wind Temp Month Day n_miss_all
#> 1      41     190  7.4   67     5   1          0
#> 2      36     118  8.0   72     5   2          0
#> 3      12     149 12.6   74     5   3          0
#> 4      18     313 11.5   62     5   4          0
#> 5      NA      NA 14.3   56     5   5          2
#> 6      28      NA 14.9   66     5   6          1
#> 7      23     299  8.6   65     5   7          0
#> 8      19      99 13.8   59     5   8          0
#> 9       8      19 20.1   61     5   9          0
#> 10     NA     194  8.6   69     5  10          1
#> 11      7      NA  6.9   74     5  11          1
#> 12     16     256  9.7   69     5  12          0
#> 13     11     290  9.2   66     5  13          0
#> 14     14     274 10.9   68     5  14          0
#> 15     18      65 13.2   58     5  15          0
#> 16     14     334 11.5   64     5  16          0
#> 17     34     307 12.0   66     5  17          0
#> 18      6      78 18.4   57     5  18          0
#> 19     30     322 11.5   68     5  19          0
#> 20     11      44  9.7   62     5  20          0
#> 21      1       8  9.7   59     5  21          0
#> 22     11     320 16.6   73     5  22          0
#> 23      4      25  9.7   61     5  23          0
#> 24     32      92 12.0   61     5  24          0
#> 25     NA      66 16.6   57     5  25          1
#> 26     NA     266 14.9   58     5  26          1
#> 27     NA      NA  8.0   57     5  27          2
#> 28     23      13 12.0   67     5  28          0
#> 29     45     252 14.9   81     5  29          0
#> 30    115     223  5.7   79     5  30          0
#> 31     37     279  7.4   76     5  31          0
#> 32     NA     286  8.6   78     6   1          1
#> 33     NA     287  9.7   74     6   2          1
#> 34     NA     242 16.1   67     6   3          1
#> 35     NA     186  9.2   84     6   4          1
#> 36     NA     220  8.6   85     6   5          1
#> 37     NA     264 14.3   79     6   6          1
#> 38     29     127  9.7   82     6   7          0
#> 39     NA     273  6.9   87     6   8          1
#> 40     71     291 13.8   90     6   9          0
#> 41     39     323 11.5   87     6  10          0
#> 42     NA     259 10.9   93     6  11          1
#> 43     NA     250  9.2   92     6  12          1
#> 44     23     148  8.0   82     6  13          0
#> 45     NA     332 13.8   80     6  14          1
#> 46     NA     322 11.5   79     6  15          1
#> 47     21     191 14.9   77     6  16          0
#> 48     37     284 20.7   72     6  17          0
#> 49     20      37  9.2   65     6  18          0
#> 50     12     120 11.5   73     6  19          0
#> 51     13     137 10.3   76     6  20          0
#> 52     NA     150  6.3   77     6  21          1
#> 53     NA      59  1.7   76     6  22          1
#> 54     NA      91  4.6   76     6  23          1
#> 55     NA     250  6.3   76     6  24          1
#> 56     NA     135  8.0   75     6  25          1
#> 57     NA     127  8.0   78     6  26          1
#> 58     NA      47 10.3   73     6  27          1
#> 59     NA      98 11.5   80     6  28          1
#> 60     NA      31 14.9   77     6  29          1
#> 61     NA     138  8.0   83     6  30          1
#> 62    135     269  4.1   84     7   1          0
#> 63     49     248  9.2   85     7   2          0
#> 64     32     236  9.2   81     7   3          0
#> 65     NA     101 10.9   84     7   4          1
#> 66     64     175  4.6   83     7   5          0
#> 67     40     314 10.9   83     7   6          0
#> 68     77     276  5.1   88     7   7          0
#> 69     97     267  6.3   92     7   8          0
#> 70     97     272  5.7   92     7   9          0
#> 71     85     175  7.4   89     7  10          0
#> 72     NA     139  8.6   82     7  11          1
#> 73     10     264 14.3   73     7  12          0
#> 74     27     175 14.9   81     7  13          0
#> 75     NA     291 14.9   91     7  14          1
#> 76      7      48 14.3   80     7  15          0
#> 77     48     260  6.9   81     7  16          0
#> 78     35     274 10.3   82     7  17          0
#> 79     61     285  6.3   84     7  18          0
#> 80     79     187  5.1   87     7  19          0
#> 81     63     220 11.5   85     7  20          0
#> 82     16       7  6.9   74     7  21          0
#> 83     NA     258  9.7   81     7  22          1
#> 84     NA     295 11.5   82     7  23          1
#> 85     80     294  8.6   86     7  24          0
#> 86    108     223  8.0   85     7  25          0
#> 87     20      81  8.6   82     7  26          0
#> 88     52      82 12.0   86     7  27          0
#> 89     82     213  7.4   88     7  28          0
#> 90     50     275  7.4   86     7  29          0
#> 91     64     253  7.4   83     7  30          0
#> 92     59     254  9.2   81     7  31          0
#> 93     39      83  6.9   81     8   1          0
#> 94      9      24 13.8   81     8   2          0
#> 95     16      77  7.4   82     8   3          0
#> 96     78      NA  6.9   86     8   4          1
#> 97     35      NA  7.4   85     8   5          1
#> 98     66      NA  4.6   87     8   6          1
#> 99    122     255  4.0   89     8   7          0
#> 100    89     229 10.3   90     8   8          0
#> 101   110     207  8.0   90     8   9          0
#> 102    NA     222  8.6   92     8  10          1
#> 103    NA     137 11.5   86     8  11          1
#> 104    44     192 11.5   86     8  12          0
#> 105    28     273 11.5   82     8  13          0
#> 106    65     157  9.7   80     8  14          0
#> 107    NA      64 11.5   79     8  15          1
#> 108    22      71 10.3   77     8  16          0
#> 109    59      51  6.3   79     8  17          0
#> 110    23     115  7.4   76     8  18          0
#> 111    31     244 10.9   78     8  19          0
#> 112    44     190 10.3   78     8  20          0
#> 113    21     259 15.5   77     8  21          0
#> 114     9      36 14.3   72     8  22          0
#> 115    NA     255 12.6   75     8  23          1
#> 116    45     212  9.7   79     8  24          0
#> 117   168     238  3.4   81     8  25          0
#> 118    73     215  8.0   86     8  26          0
#> 119    NA     153  5.7   88     8  27          1
#> 120    76     203  9.7   97     8  28          0
#> 121   118     225  2.3   94     8  29          0
#> 122    84     237  6.3   96     8  30          0
#> 123    85     188  6.3   94     8  31          0
#> 124    96     167  6.9   91     9   1          0
#> 125    78     197  5.1   92     9   2          0
#> 126    73     183  2.8   93     9   3          0
#> 127    91     189  4.6   93     9   4          0
#> 128    47      95  7.4   87     9   5          0
#> 129    32      92 15.5   84     9   6          0
#> 130    20     252 10.9   80     9   7          0
#> 131    23     220 10.3   78     9   8          0
#> 132    21     230 10.9   75     9   9          0
#> 133    24     259  9.7   73     9  10          0
#> 134    44     236 14.9   81     9  11          0
#> 135    21     259 15.5   76     9  12          0
#> 136    28     238  6.3   77     9  13          0
#> 137     9      24 10.9   71     9  14          0
#> 138    13     112 11.5   71     9  15          0
#> 139    46     237  6.9   78     9  16          0
#> 140    18     224 13.8   67     9  17          0
#> 141    13      27 10.3   76     9  18          0
#> 142    24     238 10.3   68     9  19          0
#> 143    16     201  8.0   82     9  20          0
#> 144    13     238 12.6   64     9  21          0
#> 145    23      14  9.2   71     9  22          0
#> 146    36     139 10.3   81     9  23          0
#> 147     7      49 10.3   69     9  24          0
#> 148    14      20 16.6   63     9  25          0
#> 149    30     193  6.9   70     9  26          0
#> 150    NA     145 13.2   77     9  27          1
#> 151    14     191 14.3   75     9  28          0
#> 152    18     131  8.0   76     9  29          0
#> 153    20     223 11.5   68     9  30          0
airquality %>% add_n_miss(Ozone, Solar.R)
#>     Ozone Solar.R Wind Temp Month Day n_miss_vars
#> 1      41     190  7.4   67     5   1           0
#> 2      36     118  8.0   72     5   2           0
#> 3      12     149 12.6   74     5   3           0
#> 4      18     313 11.5   62     5   4           0
#> 5      NA      NA 14.3   56     5   5           2
#> 6      28      NA 14.9   66     5   6           1
#> 7      23     299  8.6   65     5   7           0
#> 8      19      99 13.8   59     5   8           0
#> 9       8      19 20.1   61     5   9           0
#> 10     NA     194  8.6   69     5  10           1
#> 11      7      NA  6.9   74     5  11           1
#> 12     16     256  9.7   69     5  12           0
#> 13     11     290  9.2   66     5  13           0
#> 14     14     274 10.9   68     5  14           0
#> 15     18      65 13.2   58     5  15           0
#> 16     14     334 11.5   64     5  16           0
#> 17     34     307 12.0   66     5  17           0
#> 18      6      78 18.4   57     5  18           0
#> 19     30     322 11.5   68     5  19           0
#> 20     11      44  9.7   62     5  20           0
#> 21      1       8  9.7   59     5  21           0
#> 22     11     320 16.6   73     5  22           0
#> 23      4      25  9.7   61     5  23           0
#> 24     32      92 12.0   61     5  24           0
#> 25     NA      66 16.6   57     5  25           1
#> 26     NA     266 14.9   58     5  26           1
#> 27     NA      NA  8.0   57     5  27           2
#> 28     23      13 12.0   67     5  28           0
#> 29     45     252 14.9   81     5  29           0
#> 30    115     223  5.7   79     5  30           0
#> 31     37     279  7.4   76     5  31           0
#> 32     NA     286  8.6   78     6   1           1
#> 33     NA     287  9.7   74     6   2           1
#> 34     NA     242 16.1   67     6   3           1
#> 35     NA     186  9.2   84     6   4           1
#> 36     NA     220  8.6   85     6   5           1
#> 37     NA     264 14.3   79     6   6           1
#> 38     29     127  9.7   82     6   7           0
#> 39     NA     273  6.9   87     6   8           1
#> 40     71     291 13.8   90     6   9           0
#> 41     39     323 11.5   87     6  10           0
#> 42     NA     259 10.9   93     6  11           1
#> 43     NA     250  9.2   92     6  12           1
#> 44     23     148  8.0   82     6  13           0
#> 45     NA     332 13.8   80     6  14           1
#> 46     NA     322 11.5   79     6  15           1
#> 47     21     191 14.9   77     6  16           0
#> 48     37     284 20.7   72     6  17           0
#> 49     20      37  9.2   65     6  18           0
#> 50     12     120 11.5   73     6  19           0
#> 51     13     137 10.3   76     6  20           0
#> 52     NA     150  6.3   77     6  21           1
#> 53     NA      59  1.7   76     6  22           1
#> 54     NA      91  4.6   76     6  23           1
#> 55     NA     250  6.3   76     6  24           1
#> 56     NA     135  8.0   75     6  25           1
#> 57     NA     127  8.0   78     6  26           1
#> 58     NA      47 10.3   73     6  27           1
#> 59     NA      98 11.5   80     6  28           1
#> 60     NA      31 14.9   77     6  29           1
#> 61     NA     138  8.0   83     6  30           1
#> 62    135     269  4.1   84     7   1           0
#> 63     49     248  9.2   85     7   2           0
#> 64     32     236  9.2   81     7   3           0
#> 65     NA     101 10.9   84     7   4           1
#> 66     64     175  4.6   83     7   5           0
#> 67     40     314 10.9   83     7   6           0
#> 68     77     276  5.1   88     7   7           0
#> 69     97     267  6.3   92     7   8           0
#> 70     97     272  5.7   92     7   9           0
#> 71     85     175  7.4   89     7  10           0
#> 72     NA     139  8.6   82     7  11           1
#> 73     10     264 14.3   73     7  12           0
#> 74     27     175 14.9   81     7  13           0
#> 75     NA     291 14.9   91     7  14           1
#> 76      7      48 14.3   80     7  15           0
#> 77     48     260  6.9   81     7  16           0
#> 78     35     274 10.3   82     7  17           0
#> 79     61     285  6.3   84     7  18           0
#> 80     79     187  5.1   87     7  19           0
#> 81     63     220 11.5   85     7  20           0
#> 82     16       7  6.9   74     7  21           0
#> 83     NA     258  9.7   81     7  22           1
#> 84     NA     295 11.5   82     7  23           1
#> 85     80     294  8.6   86     7  24           0
#> 86    108     223  8.0   85     7  25           0
#> 87     20      81  8.6   82     7  26           0
#> 88     52      82 12.0   86     7  27           0
#> 89     82     213  7.4   88     7  28           0
#> 90     50     275  7.4   86     7  29           0
#> 91     64     253  7.4   83     7  30           0
#> 92     59     254  9.2   81     7  31           0
#> 93     39      83  6.9   81     8   1           0
#> 94      9      24 13.8   81     8   2           0
#> 95     16      77  7.4   82     8   3           0
#> 96     78      NA  6.9   86     8   4           1
#> 97     35      NA  7.4   85     8   5           1
#> 98     66      NA  4.6   87     8   6           1
#> 99    122     255  4.0   89     8   7           0
#> 100    89     229 10.3   90     8   8           0
#> 101   110     207  8.0   90     8   9           0
#> 102    NA     222  8.6   92     8  10           1
#> 103    NA     137 11.5   86     8  11           1
#> 104    44     192 11.5   86     8  12           0
#> 105    28     273 11.5   82     8  13           0
#> 106    65     157  9.7   80     8  14           0
#> 107    NA      64 11.5   79     8  15           1
#> 108    22      71 10.3   77     8  16           0
#> 109    59      51  6.3   79     8  17           0
#> 110    23     115  7.4   76     8  18           0
#> 111    31     244 10.9   78     8  19           0
#> 112    44     190 10.3   78     8  20           0
#> 113    21     259 15.5   77     8  21           0
#> 114     9      36 14.3   72     8  22           0
#> 115    NA     255 12.6   75     8  23           1
#> 116    45     212  9.7   79     8  24           0
#> 117   168     238  3.4   81     8  25           0
#> 118    73     215  8.0   86     8  26           0
#> 119    NA     153  5.7   88     8  27           1
#> 120    76     203  9.7   97     8  28           0
#> 121   118     225  2.3   94     8  29           0
#> 122    84     237  6.3   96     8  30           0
#> 123    85     188  6.3   94     8  31           0
#> 124    96     167  6.9   91     9   1           0
#> 125    78     197  5.1   92     9   2           0
#> 126    73     183  2.8   93     9   3           0
#> 127    91     189  4.6   93     9   4           0
#> 128    47      95  7.4   87     9   5           0
#> 129    32      92 15.5   84     9   6           0
#> 130    20     252 10.9   80     9   7           0
#> 131    23     220 10.3   78     9   8           0
#> 132    21     230 10.9   75     9   9           0
#> 133    24     259  9.7   73     9  10           0
#> 134    44     236 14.9   81     9  11           0
#> 135    21     259 15.5   76     9  12           0
#> 136    28     238  6.3   77     9  13           0
#> 137     9      24 10.9   71     9  14           0
#> 138    13     112 11.5   71     9  15           0
#> 139    46     237  6.9   78     9  16           0
#> 140    18     224 13.8   67     9  17           0
#> 141    13      27 10.3   76     9  18           0
#> 142    24     238 10.3   68     9  19           0
#> 143    16     201  8.0   82     9  20           0
#> 144    13     238 12.6   64     9  21           0
#> 145    23      14  9.2   71     9  22           0
#> 146    36     139 10.3   81     9  23           0
#> 147     7      49 10.3   69     9  24           0
#> 148    14      20 16.6   63     9  25           0
#> 149    30     193  6.9   70     9  26           0
#> 150    NA     145 13.2   77     9  27           1
#> 151    14     191 14.3   75     9  28           0
#> 152    18     131  8.0   76     9  29           0
#> 153    20     223 11.5   68     9  30           0
airquality %>% add_n_miss(dplyr::contains("o"))
#>     Ozone Solar.R Wind Temp Month Day n_miss_vars
#> 1      41     190  7.4   67     5   1           0
#> 2      36     118  8.0   72     5   2           0
#> 3      12     149 12.6   74     5   3           0
#> 4      18     313 11.5   62     5   4           0
#> 5      NA      NA 14.3   56     5   5           2
#> 6      28      NA 14.9   66     5   6           1
#> 7      23     299  8.6   65     5   7           0
#> 8      19      99 13.8   59     5   8           0
#> 9       8      19 20.1   61     5   9           0
#> 10     NA     194  8.6   69     5  10           1
#> 11      7      NA  6.9   74     5  11           1
#> 12     16     256  9.7   69     5  12           0
#> 13     11     290  9.2   66     5  13           0
#> 14     14     274 10.9   68     5  14           0
#> 15     18      65 13.2   58     5  15           0
#> 16     14     334 11.5   64     5  16           0
#> 17     34     307 12.0   66     5  17           0
#> 18      6      78 18.4   57     5  18           0
#> 19     30     322 11.5   68     5  19           0
#> 20     11      44  9.7   62     5  20           0
#> 21      1       8  9.7   59     5  21           0
#> 22     11     320 16.6   73     5  22           0
#> 23      4      25  9.7   61     5  23           0
#> 24     32      92 12.0   61     5  24           0
#> 25     NA      66 16.6   57     5  25           1
#> 26     NA     266 14.9   58     5  26           1
#> 27     NA      NA  8.0   57     5  27           2
#> 28     23      13 12.0   67     5  28           0
#> 29     45     252 14.9   81     5  29           0
#> 30    115     223  5.7   79     5  30           0
#> 31     37     279  7.4   76     5  31           0
#> 32     NA     286  8.6   78     6   1           1
#> 33     NA     287  9.7   74     6   2           1
#> 34     NA     242 16.1   67     6   3           1
#> 35     NA     186  9.2   84     6   4           1
#> 36     NA     220  8.6   85     6   5           1
#> 37     NA     264 14.3   79     6   6           1
#> 38     29     127  9.7   82     6   7           0
#> 39     NA     273  6.9   87     6   8           1
#> 40     71     291 13.8   90     6   9           0
#> 41     39     323 11.5   87     6  10           0
#> 42     NA     259 10.9   93     6  11           1
#> 43     NA     250  9.2   92     6  12           1
#> 44     23     148  8.0   82     6  13           0
#> 45     NA     332 13.8   80     6  14           1
#> 46     NA     322 11.5   79     6  15           1
#> 47     21     191 14.9   77     6  16           0
#> 48     37     284 20.7   72     6  17           0
#> 49     20      37  9.2   65     6  18           0
#> 50     12     120 11.5   73     6  19           0
#> 51     13     137 10.3   76     6  20           0
#> 52     NA     150  6.3   77     6  21           1
#> 53     NA      59  1.7   76     6  22           1
#> 54     NA      91  4.6   76     6  23           1
#> 55     NA     250  6.3   76     6  24           1
#> 56     NA     135  8.0   75     6  25           1
#> 57     NA     127  8.0   78     6  26           1
#> 58     NA      47 10.3   73     6  27           1
#> 59     NA      98 11.5   80     6  28           1
#> 60     NA      31 14.9   77     6  29           1
#> 61     NA     138  8.0   83     6  30           1
#> 62    135     269  4.1   84     7   1           0
#> 63     49     248  9.2   85     7   2           0
#> 64     32     236  9.2   81     7   3           0
#> 65     NA     101 10.9   84     7   4           1
#> 66     64     175  4.6   83     7   5           0
#> 67     40     314 10.9   83     7   6           0
#> 68     77     276  5.1   88     7   7           0
#> 69     97     267  6.3   92     7   8           0
#> 70     97     272  5.7   92     7   9           0
#> 71     85     175  7.4   89     7  10           0
#> 72     NA     139  8.6   82     7  11           1
#> 73     10     264 14.3   73     7  12           0
#> 74     27     175 14.9   81     7  13           0
#> 75     NA     291 14.9   91     7  14           1
#> 76      7      48 14.3   80     7  15           0
#> 77     48     260  6.9   81     7  16           0
#> 78     35     274 10.3   82     7  17           0
#> 79     61     285  6.3   84     7  18           0
#> 80     79     187  5.1   87     7  19           0
#> 81     63     220 11.5   85     7  20           0
#> 82     16       7  6.9   74     7  21           0
#> 83     NA     258  9.7   81     7  22           1
#> 84     NA     295 11.5   82     7  23           1
#> 85     80     294  8.6   86     7  24           0
#> 86    108     223  8.0   85     7  25           0
#> 87     20      81  8.6   82     7  26           0
#> 88     52      82 12.0   86     7  27           0
#> 89     82     213  7.4   88     7  28           0
#> 90     50     275  7.4   86     7  29           0
#> 91     64     253  7.4   83     7  30           0
#> 92     59     254  9.2   81     7  31           0
#> 93     39      83  6.9   81     8   1           0
#> 94      9      24 13.8   81     8   2           0
#> 95     16      77  7.4   82     8   3           0
#> 96     78      NA  6.9   86     8   4           1
#> 97     35      NA  7.4   85     8   5           1
#> 98     66      NA  4.6   87     8   6           1
#> 99    122     255  4.0   89     8   7           0
#> 100    89     229 10.3   90     8   8           0
#> 101   110     207  8.0   90     8   9           0
#> 102    NA     222  8.6   92     8  10           1
#> 103    NA     137 11.5   86     8  11           1
#> 104    44     192 11.5   86     8  12           0
#> 105    28     273 11.5   82     8  13           0
#> 106    65     157  9.7   80     8  14           0
#> 107    NA      64 11.5   79     8  15           1
#> 108    22      71 10.3   77     8  16           0
#> 109    59      51  6.3   79     8  17           0
#> 110    23     115  7.4   76     8  18           0
#> 111    31     244 10.9   78     8  19           0
#> 112    44     190 10.3   78     8  20           0
#> 113    21     259 15.5   77     8  21           0
#> 114     9      36 14.3   72     8  22           0
#> 115    NA     255 12.6   75     8  23           1
#> 116    45     212  9.7   79     8  24           0
#> 117   168     238  3.4   81     8  25           0
#> 118    73     215  8.0   86     8  26           0
#> 119    NA     153  5.7   88     8  27           1
#> 120    76     203  9.7   97     8  28           0
#> 121   118     225  2.3   94     8  29           0
#> 122    84     237  6.3   96     8  30           0
#> 123    85     188  6.3   94     8  31           0
#> 124    96     167  6.9   91     9   1           0
#> 125    78     197  5.1   92     9   2           0
#> 126    73     183  2.8   93     9   3           0
#> 127    91     189  4.6   93     9   4           0
#> 128    47      95  7.4   87     9   5           0
#> 129    32      92 15.5   84     9   6           0
#> 130    20     252 10.9   80     9   7           0
#> 131    23     220 10.3   78     9   8           0
#> 132    21     230 10.9   75     9   9           0
#> 133    24     259  9.7   73     9  10           0
#> 134    44     236 14.9   81     9  11           0
#> 135    21     259 15.5   76     9  12           0
#> 136    28     238  6.3   77     9  13           0
#> 137     9      24 10.9   71     9  14           0
#> 138    13     112 11.5   71     9  15           0
#> 139    46     237  6.9   78     9  16           0
#> 140    18     224 13.8   67     9  17           0
#> 141    13      27 10.3   76     9  18           0
#> 142    24     238 10.3   68     9  19           0
#> 143    16     201  8.0   82     9  20           0
#> 144    13     238 12.6   64     9  21           0
#> 145    23      14  9.2   71     9  22           0
#> 146    36     139 10.3   81     9  23           0
#> 147     7      49 10.3   69     9  24           0
#> 148    14      20 16.6   63     9  25           0
#> 149    30     193  6.9   70     9  26           0
#> 150    NA     145 13.2   77     9  27           1
#> 151    14     191 14.3   75     9  28           0
#> 152    18     131  8.0   76     9  29           0
#> 153    20     223 11.5   68     9  30           0