add_span_counter {naniar}	R Documentation

Add a counter variable for a span of dataframe

Description

Adds a variable, span_counter to a dataframe. Used internally to facilitate counting of missing values over a given span.

Usage

add_span_counter(data, span_size)

Arguments

data
data.frame

span_size
integer

Value

data.frame with extra variable "span_counter".

Examples

## Not run: 
add_span_counter(pedestrian, span_size = 100)

## End(Not run)

[Package naniar version 0.4.2 Index]