any-na {naniar}	R Documentation

Identify if there are any missing or complete values

Description

It is useful to search for any instances of missing or complete values. There Are two functions that do this in naniar - any_miss and it's alias any_na. These bother under the hood call anyNA. any_complete is the complement to any_miss - it returns TRUE if there are any complete values.

Usage

any_na(x)

any_miss(x)

any_complete(x)

Arguments

x
an R object to be tested

See Also

all_miss() all_complete

Examples

anyNA(airquality)
any_na(airquality)
any_miss(airquality)
any_complete(airquality)

[Package naniar version 0.4.2 Index]