Helper function to determine whether there are any missings

Helper function to determine whether there are any missings

any_row_miss(x)

Arguments

x
a vector

Value

logical vector TRUE = missing FALSE = complete