as_shadow.data.frame {naniar}	R Documentation

Create shadow data

Description

Return a tibble in shadow matrix form, where the variables are the same but have a suffix _NA attached to distinguish them.

Usage

## S3 method for class 'data.frame'
as_shadow(data, ...)

Arguments

data
dataframe

...
selected variables to use

Examples

as_shadow(airquality)

[Package naniar version 0.4.2 Index]