as_shadow_upset {naniar}	R Documentation

Convert data into shadow format for doing an upset plot

Description

Upset plots are a way of visualising common sets, this function transforms the data into a format that feeds directly into an upset plot

Usage

as_shadow_upset(data)

Arguments

data
a data.frame

Value

a data.frame

Examples

## Not run: 

library(UpSetR)
airquality %>%
  as_shadow_upset() %>%
  upset()

## End(Not run)

[Package naniar version 0.4.2 Index]