The Behavioral Risk Factor Surveillance System (BRFSS) Survey

The data is a subset of the 2009 survey from BRFSS, an ongoing data collection program designed to measure behavioral risk factors for the adult population (18 years of age or older) living in households.

data(brfss)

Source

http://www.cdc.gov/BRFSS/technical_infodata/surveydata/2009.htm

Details

Also see the codebook: http://ftp.cdc.gov/pub/data/brfss/codebook_09.rtf

Format: a data frame with 245 observations on the following 34 variables.

SEX

A factor with levels Male Female.

AGE

A numeric vector from 7 to 97.

HISPANC2

A factor with levels Yes No corresponding to the question: are you Hispanic or Latino?

VETERAN2

A factor with levels 1 2 3 4 5. The question for this variable is: Have you ever served on active duty in the United States Armed Forces, either in the regular military or in a National Guard or military reserve unit? Active duty does not include training for the Reserves or National Guard, but DOES include activation, for example, for the Persian Gulf War. And the labels are meaning: 1: Yes, now on active duty; 2: Yes, on active duty during the last 12 months, but not now; 3: Yes, on active duty in the past, but not during the last 12 months; 4: No, training for Reserves or National Guard only; 5: No, never served in the military.

MARITAL

A factor with levels Married Divorced Widowed Separated NeverMarried UnmarriedCouple.

CHILDREN

A numeric vector giving the number of children less than 18 years of age in household.

EDUCA

A factor with the education levels 1 2 3 4 5 6 as 1: Never attended school or only kindergarten; 2: Grades 1 through 8 (Elementary); 3: Grades 9 through 11 (Some high school); 4: Grade 12 or GED (High school graduate); 5: College 1 year to 3 years (Some college or technical school); 6: College 4 years or more (College graduate).

EMPLOY

A factor showing the employment status with levels 1 2 3 4 5 7 8. The labels mean -- 1: Employed for wages; 2: Self-employed; 3: Out of work for more than 1 year; 4: Out of work for less that 1 year; 5: A homemaker; 6: A student; 7: Retired; 8: Unable to work.

INCOME2

The annual household income from all sources with levels <10k 10-15k 15-20k 20-25k 25-35k 35-50k 50-75k >75k Dontknow Refused.

WEIGHT2

The weight without shoes in pounds.

HEIGHT3

The weight without shoes in inches.

PREGNANT

Whether pregnant now with two levels Yes and No.

GENHLTH

The answer to the question "in general your health is" with levels Excellent VeryGood Good Fair Poor Refused.

PHYSHLTH

The number of days during the last 30 days that the respondent's physical health was not good. -7 is for "Don't know/Not sure", and -9 is for "Refused".

MENTHLTH

The number of days during the last 30 days that the respondent's mental health was not good. -7 is for "Don't know/Not sure", and -9 is for "Refused".

POORHLTH

The number of days during the last 30 days that poor physical or mental health keep the respondent from doing usual activities, such as self-care, work, or recreation. -7 is for "Don't know/Not sure", and -9 is for "Refused".

HLTHPLAN

Whether having any kind of health care coverage, including health insurance, prepaid plans such as HMOs, or government plans such as Medicare. The answer has two levels: Yes and No.

CAREGIVE

Whether providing any such care or assistance to a friend or family member during the past month, with levels Yes and No.

QLACTLM2

Whether being limited in any way in any activities because of physical, mental, or emotional problems, with levels Yes and No.

DRNKANY4

Whether having had at least one drink of any alcoholic beverage such as beer, wine, a malt beverage or liquor during the past 30 days, with levels Yes and No.

ALCDAY4

The number of days during the past 30 days that the respondent had at least one drink of any alcoholic beverage. -7 is for "Don't know/Not sure", and -9 is for "Refused".

AVEDRNK2

The number of drinks on the average the respondent had on the days when he/she drank, during the past 30 days. -7 is for "Don't know/Not sure", and -9 is for "Refused".

SMOKE100

Whether having smoked at least 100 cigarettes in the entire life, with levels Yes and No.

SMOKDAY2

The frequency of days now smoking, with levels Everyday Somedays and NotAtAll(not at all).

STOPSMK2

Whether having stopped smoking for one day or longer during the past 12 months because the respondent was trying to quit smoking, with levels Yes and No.

LASTSMK1

A factor with levels 3 4 5 6 7 8 corresponding to the question: how long has it been since last smokeing cigarettes regularly? The labels mean: 3: Within the past 6 months (3 months but less than 6 months ago); 4: Within the past year (6 months but less than 1 year ago); 5: Within the past 5 years (1 year but less than 5 years ago); 6: Within the past 10 years (5 years but less than 10 years ago); 7: 10 years or more; 8: Never smoked regularly.

FRUIT

The number of fruit the respondent eat every year, not counting juice. -7 is for "Don't know/Not sure", and -9 is for "Refused".

GREENSAL

The number of servings of green salad the respondent eat every year. -7 is for "Don't know/Not sure", and -9 is for "Refused".

POTATOES

The number of servings of potatoes, not including french fries, fried potatoes, or potato chips, that the respondent eat every year. -7 is for "Don't know/Not sure", and -9 is for "Refused".

CARROTS

The number of carrots the respondent eat every year. -7 is for "Don't know/Not sure", and -9 is for "Refused".

VEGETABL

The number of servings of vegetables the respondent eat every year, not counting carrots, potatoes, or salad. -7 is for "Don't know/Not sure", and -9 is for "Refused".

FRUITJUI

The number of fruit juices such as orange, grapefruit, or tomato that the respondent drink every year. -7 is for "Don't know/Not sure", and -9 is for "Refused".

BMI4

Body Mass Index (BMI). Computed by WEIGHT in Kilograms/(HEIGHT in Meters * HEIGHT3 in Meters). Missing if any of WEIGHT2 or HEIGHT3 is missing.

See also

library(MissingDataGUI)