Add a shadow column to a dataset

Casting a shadow shifted column performs the equivalent pattern to data %>% select(var) %>% impute_below(). This is a convenience function that makes it easy to perform certain visualisations, in line with the principle that the user should have a way to flexibly return data formats containing information about the missing data. It forms the base building block for the functions cast_shadow_shift, and cast_shadow_shift_label. It also respects the dplyr verbs starts_with, contains, ends_with, etc. to select variables.

Usage

cast_shadow(data, ...)

Arguments

data

data.frame

...

One or more unquoted variable names separated by commas. These respect the dplyr verbs starts_with, contains, ends_with, etc.

Value

data with the added variable shifted and the suffix _NA

See also

Examples

airquality %>% cast_shadow(Ozone, Solar.R)
#> # A tibble: 153 × 4
#>    Ozone Solar.R Ozone_NA Solar.R_NA
#>    <int>   <int> <fct>    <fct>     
#>  1    41     190 !NA      !NA       
#>  2    36     118 !NA      !NA       
#>  3    12     149 !NA      !NA       
#>  4    18     313 !NA      !NA       
#>  5    NA      NA NA       NA        
#>  6    28      NA !NA      NA        
#>  7    23     299 !NA      !NA       
#>  8    19      99 !NA      !NA       
#>  9     8      19 !NA      !NA       
#> 10    NA     194 NA       !NA       
#> # ℹ 143 more rows
if (FALSE) { # \dontrun{
library(ggplot2)
library(magrittr)
airquality  %>%
  cast_shadow(Ozone,Solar.R) %>%
  ggplot(aes(x = Ozone,
             colour = Solar.R_NA)) +
        geom_density()
} # }