draw_key {naniar}	R Documentation

Key drawing functions

Description

Each Geom has an associated function that draws the key when the geom needs to be displayed in a legend. These are the options built into naniar.

Usage

draw_key_missing_point(data, params, size)

Arguments

data
A single row data frame containing the scaled aesthetics to display in this key

params
A list of additional parameters supplied to the geom.

size
Width and height of key in mm.

Value

A grid grob.

[Package naniar version 0.4.2 Index]