gather_shadow {naniar}	R Documentation

Long form representation of a shadow matrix

Description

gather_shadow is a long-form representation of binding the shadow matrix to your data, producing variables named case, variable, and missing, where missing contains the missing value representation.

Usage

gather_shadow(data)

Arguments

data
a dataframe

Value

dataframe in long, format, containing information about the missings

Examples

gather_shadow(airquality)

[Package naniar version 0.4.2 Index]