Impute nearest neighbours

Impute nearest neighbours

impute_knn(.tbl, k = 5, method = "mean")

Arguments

.tbl
data.frame with missing values

k
integer, default value 5

method

"mean" or "random" - do you want to take the mean of the k neighbours, do you want to take a random value from the k neighbours.

Value

data.frame with imputed values

Note

need to add separate functions for knn_avg and knn_rand, and also scoped variants. Ideally, this should work on a vector first, with different methods for each part

Examples

airquality %>%
  bind_shadow() %>%
  impute_knn("mean")

#> # A tibble: 153 x 12
#>    Ozone Solar.R  Wind  Temp Month   Day Ozone_NA Solar.R_NA Wind_NA Temp_NA
#>    <dbl>   <dbl> <dbl> <int> <int> <int> <fct>    <fct>      <fct>   <fct>  
#>  1  41      190    7.4    67     5     1 !NA      !NA        !NA     !NA    
#>  2  36      118    8      72     5     2 !NA      !NA        !NA     !NA    
#>  3  12      149   12.6    74     5     3 !NA      !NA        !NA     !NA    
#>  4  18      313   11.5    62     5     4 !NA      !NA        !NA     !NA    
#>  5  42.1    185.  14.3    56     5     5 NA       NA         !NA     !NA    
#>  6  28      185.  14.9    66     5     6 !NA      NA         !NA     !NA    
#>  7  23      299    8.6    65     5     7 !NA      !NA        !NA     !NA    
#>  8  19       99   13.8    59     5     8 !NA      !NA        !NA     !NA    
#>  9   8       19   20.1    61     5     9 !NA      !NA        !NA     !NA    
#> 10  42.1    194    8.6    69     5    10 NA       !NA        !NA     !NA    
#> # ... with 143 more rows, and 2 more variables: Month_NA <fct>, Day_NA <fct>

# NOT RUN {
library(ggplot2)
airquality %>%
  bind_shadow() %>%
  impute_knn() %>%
  add_label_shadow() %>%
  ggplot(aes(x = Ozone,
             y = Solar.R,
             colour = any_missing)) +
         geom_point()
# }