is_shade {naniar}	R Documentation

Detect if this is a shade

Description

This tells us if this column is a shade

Usage

is_shade(x)

are_shade(x)

any_shade(x)

Arguments

x
a vector you want to test if is a shade

Value

logical - is this a shade?

Examples

xs <- shade(c(NA, 1, 2, "3"))

is_shade(xs)
are_shade(xs)
any_shade(xs)

aq_s <- as_shadow(airquality)

is_shade(aq_s)
are_shade(aq_s)
any_shade(aq_s)
any_shade(airquality)

[Package naniar version 0.4.2 Index]