is_shadow {naniar}	R Documentation

Test if input is or are shadow variables

Description

Shadow matrix or "nabular" data is a useful way to store missing data to facilitate missing data visualisation. This data can be created using bind_shadow. is_shadow tells us if there are any shadow variables.

Usage

is_shadow(x)

is_nabular(x)

Arguments

x
a vector or data.frame

Value

logical vector of length 1

Examples

aq_sh <- as_shadow(airquality)
aq_bind <- bind_shadow(airquality)

is_shadow(aq_sh)
is_shadow(airquality)
is_shadow(aq_bind)
is_nabular(aq_bind)

[Package naniar version 0.4.2 Index]