Label a missing from one column

Label whether a value is missing in a row of one columns.

label_miss_1d(x1)

Arguments

x1
a variable of a dataframe

Value

a vector indicating whether any of these rows had missing values

Note

can we generalise label_miss to work for any number of variables?

Examples

label_miss_1d(airquality$Ozone)

#>   [1] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#>   [6] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#>  [11] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [16] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [21] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#>  [26] "Missing"     "Missing"     "Not Missing" "Not Missing" "Not Missing"
#>  [31] "Not Missing" "Missing"     "Missing"     "Missing"     "Missing"    
#>  [36] "Missing"     "Missing"     "Not Missing" "Missing"     "Not Missing"
#>  [41] "Not Missing" "Missing"     "Missing"     "Not Missing" "Missing"    
#>  [46] "Missing"     "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [51] "Not Missing" "Missing"     "Missing"     "Missing"     "Missing"    
#>  [56] "Missing"     "Missing"     "Missing"     "Missing"     "Missing"    
#>  [61] "Missing"     "Not Missing" "Not Missing" "Not Missing" "Missing"    
#>  [66] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [71] "Not Missing" "Missing"     "Not Missing" "Not Missing" "Missing"    
#>  [76] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [81] "Not Missing" "Not Missing" "Missing"     "Missing"     "Not Missing"
#>  [86] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [91] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [96] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [101] "Not Missing" "Missing"     "Missing"     "Not Missing" "Not Missing"
#> [106] "Not Missing" "Missing"     "Not Missing" "Not Missing" "Not Missing"
#> [111] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#> [116] "Not Missing" "Not Missing" "Not Missing" "Missing"     "Not Missing"
#> [121] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [126] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [131] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [136] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [141] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [146] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#> [151] "Not Missing" "Not Missing" "Not Missing"