label_shadow {naniar}	R Documentation

Label shadow values as missing or not missing

Description

Powers add_label_shadow. For the moment it is an internal function.

Usage

label_shadow(data, ..., missing = "Missing", complete = "Not Missing")

Arguments

data
data.frame

...
extra variable to label

missing
character a label for when values are missing - defaults to "Missing"

complete
character character a label for when values are complete - defaults to "Not Missing"

Value

"Missing" or "Not Missing"

[Package naniar version 0.4.2 Index]