Give NAs a more meaningful label

Returns a binary factor of !NA and NA, where !NA indicates a datum that is not missing, and NA indicates missingness. This function is what powers the factor levels in as_shadow().

label_shadow_matrix(x)

Arguments

x
a vector

Value

a vector of factors containing the labels "!NA" for Not missing and "NA" for missing.

See also

Examples

# NOT RUN {
label_shadow_matrix(airquality$Ozone)
# }