Proportion of variables containing missings or complete values

Calculate the proportion of variables that contain a single missing or complete values.

miss_var_prop(data)

complete_var_prop(data)

Arguments

data
a dataframe

Value

numeric the proportion of variables that contain missing or complete data

See also

Examples

miss_var_prop(riskfactors)

#> [1] 0.7058824

miss_var_prop(oceanbuoys)

#> [1] 0.375

complete_var_prop(riskfactors)

#> [1] 0.2941176

complete_var_prop(oceanbuoys)

#> [1] 0.625