Proportions of missings in data, variables, and cases.

Return missing data info about the dataframe, the variables, and the cases. Specifically, returning how many elements in a dataframe contain a missing value, how many elements in a variable contain a missing value, and how many elements in a case contain a missing.

miss_prop_summary(data)

Arguments

data
a dataframe

Value

a dataframe

See also

Examples

miss_prop_summary(airquality)

#> # A tibble: 1 x 3
#>       df   var  case
#>    <dbl> <dbl> <dbl>
#> 1 0.0479 0.333 0.275

library(dplyr)
airquality %>% group_by(Month) %>% miss_prop_summary()

#> # A tibble: 5 x 4
#>   Month      df   var   case
#>   <int>   <dbl> <dbl>  <dbl>
#> 1     5 0.0581    0.4 0.226 
#> 2     6 0.14      0.2 0.7   
#> 3     7 0.0323    0.2 0.161 
#> 4     8 0.0516    0.4 0.258 
#> 5     9 0.00667   0.2 0.0333