n-var-case-miss {naniar}	R Documentation

The number of variables or cases with missing values

Description

This function calculates the number of variables or cases that contain a missing value

Usage

n_var_miss(data)

n_case_miss(data)

Arguments

data
data.frame

Value

integer, number of missings

See Also

n_var_complete()

Examples

# how many variables contain missing values?
n_var_miss(airquality)
n_case_miss(airquality)

[Package naniar version 0.4.2 Index]