naniar: Data Structures, Summaries, and Visualisations for Missing Data

naniar is a package to make it easier to summarise and handle missing values in R. It strives to do this in a way that is as consistent with tidyverse principles as possible. The work is fully discussed at Tierney & Cook (2023) doi:10.18637/jss.v105.i07.

See also

Author

Maintainer: Nicholas Tierney nicholas.tierney@gmail.com (ORCID)

Authors:

Di Cook dicook@monash.edu (ORCID)

Miles McBain miles.mcbain@gmail.com (ORCID)

Colin Fay contact@colinfay.me (ORCID)

Other contributors:

Mitchell O'Hara-Wild [contributor]

Jim Hester james.f.hester@gmail.com [contributor]

Luke Smith [contributor]

Andrew Heiss andrew@andrewheiss.com (ORCID) [contributor]