new_shade {naniar}	R Documentation

Create a new shade factor

Description

Create a new shade factor

Usage

new_shade(x, extra_levels = NULL)

Arguments

x
a factor to convert into a shade object

extra_levels
the extra levels to give to shade objects, such as "broken_machine" and so on, which get converted into "NA_broken_machine".

Value

a new shade, which is built upon a factor

[Package naniar version 0.4.2 Index]