shade {naniar}	R Documentation

Create new levels of missing

Description

Returns (at least) factors of !NA and NA, where !NA indicates a datum that is not missing, and NA indicates missingness. It also allows you to specify some new missings, if you like. This function is what powers the factor levels in as_shadow().

Usage

shade(x, ..., extra_levels = NULL)

Arguments

x
a vector

...
additional levels of missing to add

extra_levels
is a

Examples

df <- tibble::tribble(
  ~wind, ~temp,
  -99,    45,
  68,    NA,
  72,    25
  )

shade(df$wind)

shade(df$wind,
      inst_fail = -99)

shade(df$wind,
      inst_fail = 100)

[Package naniar version 0.4.2 Index]