Which elements contain missings?

Internal function that creates a matrix containing the location of missing values in a dataframe. This may be used in the future impl_df class.

which_na(x)

Arguments

x
a dataframe

Value

a matrix with columns "row" and "col", which refer to the row and column that identify the position of a missing value in a dataframe

See also

where_na

Examples

naniar:::which_na(airquality)

#>  [1]   5  10  25  26  27  32  33  34  35  36  37  39  42  43  45  46  52  53  54
#> [20]  55  56  57  58  59  60  61  65  72  75  83  84 102 103 107 115 119 150 158
#> [39] 159 164 180 249 250 251