add_label_missings {naniar}	R Documentation

Add a column describing if there are any missings in the dataset

Description

Add a column describing if there are any missings in the dataset

Usage

add_label_missings(data, ..., missing = "Missing",
  complete = "Not Missing")

Arguments

data
data.frame

...
extra variable to label

missing
character a label for when values are missing - defaults to "Missing"

complete
character character a label for when values are complete - defaults to "Not Missing"

Value

data.frame with a column "any_missing" that is either "Not Missing" or "Missing" for the purposes of plotting / exploration / nice print methods

See Also

bind_shadow() add_any_miss() add_label_missings() add_label_shadow() add_miss_cluster() add_n_miss() add_prop_miss() add_shadow_shift() cast_shadow()

Examples

airquality %>% add_label_missings()
airquality %>% add_label_missings(Ozone)
airquality %>% add_label_missings(Ozone, Solar.R)
airquality %>% add_label_missings(Ozone, Solar.R, missing = "yes", complete = "no")

[Package naniar version 0.4.2 Index]