any_row_miss {naniar}	R Documentation

Helper function to determine whether there are any missings

Description

Helper function to determine whether there are any missings

Usage

any_row_miss(x)

Arguments

x
a vector

Value

logical vector TRUE = missing FALSE = complete

[Package naniar version 0.4.2 Index]