as_shadow {naniar}	R Documentation

Create shadows

Description

Representing missing data structure is achieved using the shadow matrix, introduced in Swayne and Buja. The shadow matrix is the same dimension as the data, and consists of binary indicators of missingness of data values, where missing is represented as "NA", and not missing is represented as "!NA". Although these may be represented as 1 and 0, respectively.

Usage

as_shadow(data, ...)

Arguments

data
dataframe

...
selected variables to use

Value

appended shadow with column names

[Package naniar version 0.4.2 Index]