gg_miss_upset {naniar}	R Documentation

Plot the pattern of missingness using an upset plot.

Description

Upset plots are a way of visualising common sets, gg_miss_upset`` shows the number of missing values for each of the sets of data. The default option of gg_miss_upsetis taken fromUpSetR::upset- which is to use up to 5 sets and up to 40 interactions. We also set the ordering to by the frequency of the intersections. Settingnsets = 5means to look at 5 variables and their combinations. The number of combinations or ratherintersectionsis controlled bynintersects. If there are 40 intersections, there will be 40 combinations of variables explored. The number of sets and intersections can be changed by passing arguments nsets = 10to look at 10 sets of variables, andnintersects = 50' to look at 50 intersections.

Usage

gg_miss_upset(data, order.by = "freq", ...)

Arguments

data
data.frame

order.by
(from UpSetR::upset) How the intersections in the matrix should be ordered by. Options include frequency (entered as "freq"), degree, or both in any order. See ?UpSetR::upset for more options

...
arguments to pass to upset plot - see ?UpSetR::upset

Value

a ggplot visualisation of missing data

Examples

## Not run: 
gg_miss_upset(airquality)
gg_miss_upset(pedestrian)
gg_miss_upset(riskfactors)
gg_miss_upset(riskfactors, nsets = 10)
gg_miss_upset(riskfactors, nsets = 10, nintersects = 10)

## End(Not run)

[Package naniar version 0.4.2 Index]