impute_below {naniar}	R Documentation

Impute data with values shifted 10% below range.

Description

It can be useful in exploratory graphics to impute data outside the range of the data. impute_below imputes all variables with missings to have values 10 values adds a new string or label. It is powered by shadow_shift, so please see the documentation for shadow_shift() to full details on the different implementations.

Usage

impute_below(...)

Arguments

...
extra arguments to pass - see shadow_shift() for discussion on this.

[Package naniar version 0.4.2 Index]