Impute data with values shifted 10 percent below range.

It can be useful in exploratory graphics to impute data outside the range of the data. impute_below imputes all variables with missings to have values 10 percent below the range for numeric values, and for character or factor values adds a new string or label. It is powered by shadow_shift, so please see the documentation for shadow_shift() to full details on the different implementations.

impute_below(...)

Arguments

...
extra arguments to pass - see shadow_shift() for discussion on this.