Is there a missing value in the row of a dataframe?

Creates a character vector describing presence/absense of missing values

label_missings(data)

Arguments

data
a dataframe or set of vectors of the same length

Value

character vector of "Missing" and "Not Missing".

Examples

label_missings(airquality)

#>   [1] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#>   [6] "Missing"     "Not Missing" "Not Missing" "Not Missing" "Missing"    
#>  [11] "Missing"     "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [16] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [21] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#>  [26] "Missing"     "Missing"     "Not Missing" "Not Missing" "Not Missing"
#>  [31] "Not Missing" "Missing"     "Missing"     "Missing"     "Missing"    
#>  [36] "Missing"     "Missing"     "Not Missing" "Missing"     "Not Missing"
#>  [41] "Not Missing" "Missing"     "Missing"     "Not Missing" "Missing"    
#>  [46] "Missing"     "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [51] "Not Missing" "Missing"     "Missing"     "Missing"     "Missing"    
#>  [56] "Missing"     "Missing"     "Missing"     "Missing"     "Missing"    
#>  [61] "Missing"     "Not Missing" "Not Missing" "Not Missing" "Missing"    
#>  [66] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [71] "Not Missing" "Missing"     "Not Missing" "Not Missing" "Missing"    
#>  [76] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [81] "Not Missing" "Not Missing" "Missing"     "Missing"     "Not Missing"
#>  [86] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [91] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#>  [96] "Missing"     "Missing"     "Missing"     "Not Missing" "Not Missing"
#> [101] "Not Missing" "Missing"     "Missing"     "Not Missing" "Not Missing"
#> [106] "Not Missing" "Missing"     "Not Missing" "Not Missing" "Not Missing"
#> [111] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#> [116] "Not Missing" "Not Missing" "Not Missing" "Missing"     "Not Missing"
#> [121] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [126] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [131] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [136] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [141] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Not Missing"
#> [146] "Not Missing" "Not Missing" "Not Missing" "Not Missing" "Missing"    
#> [151] "Not Missing" "Not Missing" "Not Missing"

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

airquality %>% mutate(is_missing = label_missings(airquality))

#>     Ozone Solar.R Wind Temp Month Day  is_missing
#> 1      41     190  7.4   67     5   1 Not Missing
#> 2      36     118  8.0   72     5   2 Not Missing
#> 3      12     149 12.6   74     5   3 Not Missing
#> 4      18     313 11.5   62     5   4 Not Missing
#> 5      NA      NA 14.3   56     5   5     Missing
#> 6      28      NA 14.9   66     5   6     Missing
#> 7      23     299  8.6   65     5   7 Not Missing
#> 8      19      99 13.8   59     5   8 Not Missing
#> 9       8      19 20.1   61     5   9 Not Missing
#> 10     NA     194  8.6   69     5  10     Missing
#> 11      7      NA  6.9   74     5  11     Missing
#> 12     16     256  9.7   69     5  12 Not Missing
#> 13     11     290  9.2   66     5  13 Not Missing
#> 14     14     274 10.9   68     5  14 Not Missing
#> 15     18      65 13.2   58     5  15 Not Missing
#> 16     14     334 11.5   64     5  16 Not Missing
#> 17     34     307 12.0   66     5  17 Not Missing
#> 18      6      78 18.4   57     5  18 Not Missing
#> 19     30     322 11.5   68     5  19 Not Missing
#> 20     11      44  9.7   62     5  20 Not Missing
#> 21      1       8  9.7   59     5  21 Not Missing
#> 22     11     320 16.6   73     5  22 Not Missing
#> 23      4      25  9.7   61     5  23 Not Missing
#> 24     32      92 12.0   61     5  24 Not Missing
#> 25     NA      66 16.6   57     5  25     Missing
#> 26     NA     266 14.9   58     5  26     Missing
#> 27     NA      NA  8.0   57     5  27     Missing
#> 28     23      13 12.0   67     5  28 Not Missing
#> 29     45     252 14.9   81     5  29 Not Missing
#> 30    115     223  5.7   79     5  30 Not Missing
#> 31     37     279  7.4   76     5  31 Not Missing
#> 32     NA     286  8.6   78     6   1     Missing
#> 33     NA     287  9.7   74     6   2     Missing
#> 34     NA     242 16.1   67     6   3     Missing
#> 35     NA     186  9.2   84     6   4     Missing
#> 36     NA     220  8.6   85     6   5     Missing
#> 37     NA     264 14.3   79     6   6     Missing
#> 38     29     127  9.7   82     6   7 Not Missing
#> 39     NA     273  6.9   87     6   8     Missing
#> 40     71     291 13.8   90     6   9 Not Missing
#> 41     39     323 11.5   87     6  10 Not Missing
#> 42     NA     259 10.9   93     6  11     Missing
#> 43     NA     250  9.2   92     6  12     Missing
#> 44     23     148  8.0   82     6  13 Not Missing
#> 45     NA     332 13.8   80     6  14     Missing
#> 46     NA     322 11.5   79     6  15     Missing
#> 47     21     191 14.9   77     6  16 Not Missing
#> 48     37     284 20.7   72     6  17 Not Missing
#> 49     20      37  9.2   65     6  18 Not Missing
#> 50     12     120 11.5   73     6  19 Not Missing
#> 51     13     137 10.3   76     6  20 Not Missing
#> 52     NA     150  6.3   77     6  21     Missing
#> 53     NA      59  1.7   76     6  22     Missing
#> 54     NA      91  4.6   76     6  23     Missing
#> 55     NA     250  6.3   76     6  24     Missing
#> 56     NA     135  8.0   75     6  25     Missing
#> 57     NA     127  8.0   78     6  26     Missing
#> 58     NA      47 10.3   73     6  27     Missing
#> 59     NA      98 11.5   80     6  28     Missing
#> 60     NA      31 14.9   77     6  29     Missing
#> 61     NA     138  8.0   83     6  30     Missing
#> 62    135     269  4.1   84     7   1 Not Missing
#> 63     49     248  9.2   85     7   2 Not Missing
#> 64     32     236  9.2   81     7   3 Not Missing
#> 65     NA     101 10.9   84     7   4     Missing
#> 66     64     175  4.6   83     7   5 Not Missing
#> 67     40     314 10.9   83     7   6 Not Missing
#> 68     77     276  5.1   88     7   7 Not Missing
#> 69     97     267  6.3   92     7   8 Not Missing
#> 70     97     272  5.7   92     7   9 Not Missing
#> 71     85     175  7.4   89     7  10 Not Missing
#> 72     NA     139  8.6   82     7  11     Missing
#> 73     10     264 14.3   73     7  12 Not Missing
#> 74     27     175 14.9   81     7  13 Not Missing
#> 75     NA     291 14.9   91     7  14     Missing
#> 76      7      48 14.3   80     7  15 Not Missing
#> 77     48     260  6.9   81     7  16 Not Missing
#> 78     35     274 10.3   82     7  17 Not Missing
#> 79     61     285  6.3   84     7  18 Not Missing
#> 80     79     187  5.1   87     7  19 Not Missing
#> 81     63     220 11.5   85     7  20 Not Missing
#> 82     16       7  6.9   74     7  21 Not Missing
#> 83     NA     258  9.7   81     7  22     Missing
#> 84     NA     295 11.5   82     7  23     Missing
#> 85     80     294  8.6   86     7  24 Not Missing
#> 86    108     223  8.0   85     7  25 Not Missing
#> 87     20      81  8.6   82     7  26 Not Missing
#> 88     52      82 12.0   86     7  27 Not Missing
#> 89     82     213  7.4   88     7  28 Not Missing
#> 90     50     275  7.4   86     7  29 Not Missing
#> 91     64     253  7.4   83     7  30 Not Missing
#> 92     59     254  9.2   81     7  31 Not Missing
#> 93     39      83  6.9   81     8   1 Not Missing
#> 94      9      24 13.8   81     8   2 Not Missing
#> 95     16      77  7.4   82     8   3 Not Missing
#> 96     78      NA  6.9   86     8   4     Missing
#> 97     35      NA  7.4   85     8   5     Missing
#> 98     66      NA  4.6   87     8   6     Missing
#> 99    122     255  4.0   89     8   7 Not Missing
#> 100    89     229 10.3   90     8   8 Not Missing
#> 101   110     207  8.0   90     8   9 Not Missing
#> 102    NA     222  8.6   92     8  10     Missing
#> 103    NA     137 11.5   86     8  11     Missing
#> 104    44     192 11.5   86     8  12 Not Missing
#> 105    28     273 11.5   82     8  13 Not Missing
#> 106    65     157  9.7   80     8  14 Not Missing
#> 107    NA      64 11.5   79     8  15     Missing
#> 108    22      71 10.3   77     8  16 Not Missing
#> 109    59      51  6.3   79     8  17 Not Missing
#> 110    23     115  7.4   76     8  18 Not Missing
#> 111    31     244 10.9   78     8  19 Not Missing
#> 112    44     190 10.3   78     8  20 Not Missing
#> 113    21     259 15.5   77     8  21 Not Missing
#> 114     9      36 14.3   72     8  22 Not Missing
#> 115    NA     255 12.6   75     8  23     Missing
#> 116    45     212  9.7   79     8  24 Not Missing
#> 117   168     238  3.4   81     8  25 Not Missing
#> 118    73     215  8.0   86     8  26 Not Missing
#> 119    NA     153  5.7   88     8  27     Missing
#> 120    76     203  9.7   97     8  28 Not Missing
#> 121   118     225  2.3   94     8  29 Not Missing
#> 122    84     237  6.3   96     8  30 Not Missing
#> 123    85     188  6.3   94     8  31 Not Missing
#> 124    96     167  6.9   91     9   1 Not Missing
#> 125    78     197  5.1   92     9   2 Not Missing
#> 126    73     183  2.8   93     9   3 Not Missing
#> 127    91     189  4.6   93     9   4 Not Missing
#> 128    47      95  7.4   87     9   5 Not Missing
#> 129    32      92 15.5   84     9   6 Not Missing
#> 130    20     252 10.9   80     9   7 Not Missing
#> 131    23     220 10.3   78     9   8 Not Missing
#> 132    21     230 10.9   75     9   9 Not Missing
#> 133    24     259  9.7   73     9  10 Not Missing
#> 134    44     236 14.9   81     9  11 Not Missing
#> 135    21     259 15.5   76     9  12 Not Missing
#> 136    28     238  6.3   77     9  13 Not Missing
#> 137     9      24 10.9   71     9  14 Not Missing
#> 138    13     112 11.5   71     9  15 Not Missing
#> 139    46     237  6.9   78     9  16 Not Missing
#> 140    18     224 13.8   67     9  17 Not Missing
#> 141    13      27 10.3   76     9  18 Not Missing
#> 142    24     238 10.3   68     9  19 Not Missing
#>  [ reached getOption("max.print") -- omitted 11 rows ]