Shift missing values to facilite missing data exploration/visualisation

`shadow_shift` is a window function that transforms missing values to facilitate visualisation. This has different behaviour for different types of variables. For numeric variables, the values are shifted to 10 below the minimum value for a given variable plus some jittered noise, to separate repeated values, so that missing values can be visualised along with the rest of the data. There are different shifts that can take place in different kinds of variables, which are currently under development.

shadow_shift(x, ...)

Arguments

x
a variable of interest to shift

...
extra arguments to pass

Examples

airquality$Ozone

#>   [1]  41  36  12  18  NA  28  23  19   8  NA   7  16  11  14  18  14  34   6
#>  [19]  30  11   1  11   4  32  NA  NA  NA  23  45 115  37  NA  NA  NA  NA  NA
#>  [37]  NA  29  NA  71  39  NA  NA  23  NA  NA  21  37  20  12  13  NA  NA  NA
#>  [55]  NA  NA  NA  NA  NA  NA  NA 135  49  32  NA  64  40  77  97  97  85  NA
#>  [73]  10  27  NA   7  48  35  61  79  63  16  NA  NA  80 108  20  52  82  50
#>  [91]  64  59  39   9  16  78  35  66 122  89 110  NA  NA  44  28  65  NA  22
#> [109]  59  23  31  44  21   9  NA  45 168  73  NA  76 118  84  85  96  78  73
#> [127]  91  47  32  20  23  21  24  44  21  28   9  13  46  18  13  24  16  13
#> [145]  23  36   7  14  30  NA  14  18  20

shadow_shift(airquality$Ozone)

#>   [1]  41.00000  36.00000  12.00000  18.00000 -19.72321  28.00000  23.00000
#>   [8]  19.00000   8.00000 -18.51277   7.00000  16.00000  11.00000  14.00000
#>  [15]  18.00000  14.00000  34.00000   6.00000  30.00000  11.00000   1.00000
#>  [22]  11.00000   4.00000  32.00000 -17.81863 -19.43853 -15.14310  23.00000
#>  [29]  45.00000 115.00000  37.00000 -16.17315 -14.65883 -17.85609 -13.29299
#>  [36] -16.16323 -19.60935  29.00000 -19.65780  71.00000  39.00000 -13.40961
#>  [43] -13.53728  23.00000 -19.65993 -16.48342  21.00000  37.00000  20.00000
#>  [50]  12.00000  13.00000 -17.17718 -16.74073 -13.65786 -16.78786 -12.30098
#>  [57] -13.33171 -16.77414 -17.08225 -15.98818 -19.17558 135.00000  49.00000
#>  [64]  32.00000 -14.27138  64.00000  40.00000  77.00000  97.00000  97.00000
#>  [71]  85.00000 -13.51764  10.00000  27.00000 -13.48998   7.00000  48.00000
#>  [78]  35.00000  61.00000  79.00000  63.00000  16.00000 -16.92150 -16.60335
#>  [85]  80.00000 108.00000  20.00000  52.00000  82.00000  50.00000  64.00000
#>  [92]  59.00000  39.00000   9.00000  16.00000  78.00000  35.00000  66.00000
#>  [99] 122.00000  89.00000 110.00000 -14.78907 -16.19151  44.00000  28.00000
#> [106]  65.00000 -19.73591  22.00000  59.00000  23.00000  31.00000  44.00000
#> [113]  21.00000   9.00000 -18.92235  45.00000 168.00000  73.00000 -14.86296
#> [120]  76.00000 118.00000  84.00000  85.00000  96.00000  78.00000  73.00000
#> [127]  91.00000  47.00000  32.00000  20.00000  23.00000  21.00000  24.00000
#> [134]  44.00000  21.00000  28.00000   9.00000  13.00000  46.00000  18.00000
#> [141]  13.00000  24.00000  16.00000  13.00000  23.00000  36.00000   7.00000
#> [148]  14.00000  30.00000 -14.83089  14.00000  18.00000  20.00000

library(dplyr)
airquality %>%
    mutate(Ozone_shift = shadow_shift(Ozone))

#>     Ozone Solar.R Wind Temp Month Day Ozone_shift
#> 1      41     190  7.4   67     5   1    41.00000
#> 2      36     118  8.0   72     5   2    36.00000
#> 3      12     149 12.6   74     5   3    12.00000
#> 4      18     313 11.5   62     5   4    18.00000
#> 5      NA      NA 14.3   56     5   5   -19.72321
#> 6      28      NA 14.9   66     5   6    28.00000
#> 7      23     299  8.6   65     5   7    23.00000
#> 8      19      99 13.8   59     5   8    19.00000
#> 9       8      19 20.1   61     5   9     8.00000
#> 10     NA     194  8.6   69     5  10   -18.51277
#> 11      7      NA  6.9   74     5  11     7.00000
#> 12     16     256  9.7   69     5  12    16.00000
#> 13     11     290  9.2   66     5  13    11.00000
#> 14     14     274 10.9   68     5  14    14.00000
#> 15     18      65 13.2   58     5  15    18.00000
#> 16     14     334 11.5   64     5  16    14.00000
#> 17     34     307 12.0   66     5  17    34.00000
#> 18      6      78 18.4   57     5  18     6.00000
#> 19     30     322 11.5   68     5  19    30.00000
#> 20     11      44  9.7   62     5  20    11.00000
#> 21      1       8  9.7   59     5  21     1.00000
#> 22     11     320 16.6   73     5  22    11.00000
#> 23      4      25  9.7   61     5  23     4.00000
#> 24     32      92 12.0   61     5  24    32.00000
#> 25     NA      66 16.6   57     5  25   -17.81863
#> 26     NA     266 14.9   58     5  26   -19.43853
#> 27     NA      NA  8.0   57     5  27   -15.14310
#> 28     23      13 12.0   67     5  28    23.00000
#> 29     45     252 14.9   81     5  29    45.00000
#> 30    115     223  5.7   79     5  30   115.00000
#> 31     37     279  7.4   76     5  31    37.00000
#> 32     NA     286  8.6   78     6   1   -16.17315
#> 33     NA     287  9.7   74     6   2   -14.65883
#> 34     NA     242 16.1   67     6   3   -17.85609
#> 35     NA     186  9.2   84     6   4   -13.29299
#> 36     NA     220  8.6   85     6   5   -16.16323
#> 37     NA     264 14.3   79     6   6   -19.60935
#> 38     29     127  9.7   82     6   7    29.00000
#> 39     NA     273  6.9   87     6   8   -19.65780
#> 40     71     291 13.8   90     6   9    71.00000
#> 41     39     323 11.5   87     6  10    39.00000
#> 42     NA     259 10.9   93     6  11   -13.40961
#> 43     NA     250  9.2   92     6  12   -13.53728
#> 44     23     148  8.0   82     6  13    23.00000
#> 45     NA     332 13.8   80     6  14   -19.65993
#> 46     NA     322 11.5   79     6  15   -16.48342
#> 47     21     191 14.9   77     6  16    21.00000
#> 48     37     284 20.7   72     6  17    37.00000
#> 49     20      37  9.2   65     6  18    20.00000
#> 50     12     120 11.5   73     6  19    12.00000
#> 51     13     137 10.3   76     6  20    13.00000
#> 52     NA     150  6.3   77     6  21   -17.17718
#> 53     NA      59  1.7   76     6  22   -16.74073
#> 54     NA      91  4.6   76     6  23   -13.65786
#> 55     NA     250  6.3   76     6  24   -16.78786
#> 56     NA     135  8.0   75     6  25   -12.30098
#> 57     NA     127  8.0   78     6  26   -13.33171
#> 58     NA      47 10.3   73     6  27   -16.77414
#> 59     NA      98 11.5   80     6  28   -17.08225
#> 60     NA      31 14.9   77     6  29   -15.98818
#> 61     NA     138  8.0   83     6  30   -19.17558
#> 62    135     269  4.1   84     7   1   135.00000
#> 63     49     248  9.2   85     7   2    49.00000
#> 64     32     236  9.2   81     7   3    32.00000
#> 65     NA     101 10.9   84     7   4   -14.27138
#> 66     64     175  4.6   83     7   5    64.00000
#> 67     40     314 10.9   83     7   6    40.00000
#> 68     77     276  5.1   88     7   7    77.00000
#> 69     97     267  6.3   92     7   8    97.00000
#> 70     97     272  5.7   92     7   9    97.00000
#> 71     85     175  7.4   89     7  10    85.00000
#> 72     NA     139  8.6   82     7  11   -13.51764
#> 73     10     264 14.3   73     7  12    10.00000
#> 74     27     175 14.9   81     7  13    27.00000
#> 75     NA     291 14.9   91     7  14   -13.48998
#> 76      7      48 14.3   80     7  15     7.00000
#> 77     48     260  6.9   81     7  16    48.00000
#> 78     35     274 10.3   82     7  17    35.00000
#> 79     61     285  6.3   84     7  18    61.00000
#> 80     79     187  5.1   87     7  19    79.00000
#> 81     63     220 11.5   85     7  20    63.00000
#> 82     16       7  6.9   74     7  21    16.00000
#> 83     NA     258  9.7   81     7  22   -16.92150
#> 84     NA     295 11.5   82     7  23   -16.60335
#> 85     80     294  8.6   86     7  24    80.00000
#> 86    108     223  8.0   85     7  25   108.00000
#> 87     20      81  8.6   82     7  26    20.00000
#> 88     52      82 12.0   86     7  27    52.00000
#> 89     82     213  7.4   88     7  28    82.00000
#> 90     50     275  7.4   86     7  29    50.00000
#> 91     64     253  7.4   83     7  30    64.00000
#> 92     59     254  9.2   81     7  31    59.00000
#> 93     39      83  6.9   81     8   1    39.00000
#> 94      9      24 13.8   81     8   2     9.00000
#> 95     16      77  7.4   82     8   3    16.00000
#> 96     78      NA  6.9   86     8   4    78.00000
#> 97     35      NA  7.4   85     8   5    35.00000
#> 98     66      NA  4.6   87     8   6    66.00000
#> 99    122     255  4.0   89     8   7   122.00000
#> 100    89     229 10.3   90     8   8    89.00000
#> 101   110     207  8.0   90     8   9   110.00000
#> 102    NA     222  8.6   92     8  10   -14.78907
#> 103    NA     137 11.5   86     8  11   -16.19151
#> 104    44     192 11.5   86     8  12    44.00000
#> 105    28     273 11.5   82     8  13    28.00000
#> 106    65     157  9.7   80     8  14    65.00000
#> 107    NA      64 11.5   79     8  15   -19.73591
#> 108    22      71 10.3   77     8  16    22.00000
#> 109    59      51  6.3   79     8  17    59.00000
#> 110    23     115  7.4   76     8  18    23.00000
#> 111    31     244 10.9   78     8  19    31.00000
#> 112    44     190 10.3   78     8  20    44.00000
#> 113    21     259 15.5   77     8  21    21.00000
#> 114     9      36 14.3   72     8  22     9.00000
#> 115    NA     255 12.6   75     8  23   -18.92235
#> 116    45     212  9.7   79     8  24    45.00000
#> 117   168     238  3.4   81     8  25   168.00000
#> 118    73     215  8.0   86     8  26    73.00000
#> 119    NA     153  5.7   88     8  27   -14.86296
#> 120    76     203  9.7   97     8  28    76.00000
#> 121   118     225  2.3   94     8  29   118.00000
#> 122    84     237  6.3   96     8  30    84.00000
#> 123    85     188  6.3   94     8  31    85.00000
#> 124    96     167  6.9   91     9   1    96.00000
#> 125    78     197  5.1   92     9   2    78.00000
#> 126    73     183  2.8   93     9   3    73.00000
#> 127    91     189  4.6   93     9   4    91.00000
#> 128    47      95  7.4   87     9   5    47.00000
#> 129    32      92 15.5   84     9   6    32.00000
#> 130    20     252 10.9   80     9   7    20.00000
#> 131    23     220 10.3   78     9   8    23.00000
#> 132    21     230 10.9   75     9   9    21.00000
#> 133    24     259  9.7   73     9  10    24.00000
#> 134    44     236 14.9   81     9  11    44.00000
#> 135    21     259 15.5   76     9  12    21.00000
#> 136    28     238  6.3   77     9  13    28.00000
#> 137     9      24 10.9   71     9  14     9.00000
#> 138    13     112 11.5   71     9  15    13.00000
#> 139    46     237  6.9   78     9  16    46.00000
#> 140    18     224 13.8   67     9  17    18.00000
#> 141    13      27 10.3   76     9  18    13.00000
#> 142    24     238 10.3   68     9  19    24.00000
#>  [ reached getOption("max.print") -- omitted 11 rows ]