where {naniar}	R Documentation

Split a call into two components with a useful verb name

Description

This function is used inside recode_shadow to help evaluate the formula call effectively. .where is a special function designed for use in recode_shadow, and you shouldn't use it outside of it

Usage

.where(...)

Arguments

...
case_when style formula

Value

a list of "condition" and "suffix" arguments

Examples

## Not run: 
df <- tibble::tribble(
~wind, ~temp,
-99,    45,
68,    NA,
72,    25
)

dfs <- bind_shadow(df)

recode_shadow(dfs,
              temp = .where(wind == -99 ~ "bananas"))

## End(Not run)

[Package naniar version 0.4.2 Index]